# lib/api-zod/src/generated/types/complaint.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Smart Bharat AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ComplaintStatus } from './complaintStatus';
import type { TimelineEvent } from './timelineEvent';

export interface Complaint {
  id: string;
  complaintId: string;
  userId: string;
  category: string;
  description: string;
  status: ComplaintStatus;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  priority?: string | null;
  /** @nullable */
  severity?: string | null;
  /** @nullable */
  officerName?: string | null;
  createdAt: string;
  updatedAt?: string;
  timeline?: TimelineEvent[];
}