# lib/api-zod/src/generated/types/documentInfoInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Smart Bharat AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentInfoInput {
  documentType: string;
}